# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CertificateResponse": {
        "description": "One certificate, without any key material.",
        "properties": {
          "autoRenew": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "daysRemaining": {
            "format": "int64",
            "type": "integer"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "fingerprintSha256": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "inUse": {
            "type": "boolean"
          },
          "issuer": {
            "type": "string"
          },
          "lastError": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "notBefore": {
            "format": "date-time",
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "renewBeforeDays": {
            "format": "int32",
            "type": "integer"
          },
          "revokedAt": {
            "format": "date-time",
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "autoRenew",
          "createdAt",
          "createdBy",
          "domains",
          "id",
          "inUse",
          "name",
          "provider",
          "renewBeforeDays",
          "status",
          "version"
        ],
        "type": "object"
      },
      "CloneRequest": {
        "properties": {
          "domain": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "domain",
          "name"
        ],
        "type": "object"
      },
...
      "HeaderDto": {
        "description": "A header rule.",
        "properties": {
          "direction": {
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "direction",
          "name",
          "value"
        ],
        "type": "object"
      },
      "LocationDto": {
        "description": "A location rule.",
        "properties": {
          "matchType": {
            "type": "string"
          },
          "pathPattern": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "matchType",
          "pathPattern"
        ],
        "type": "object"
      },
...
      "PageResponseProxySiteSummaryResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/ProxySiteSummaryResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ProxySiteRequest": {
        "properties": {
          "activeFrom": {
            "format": "date-time",
            "type": "string"
          },
          "connectTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "domain": {
            "maxLength": 253,
            "minLength": 0,
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "forceHttps": {
            "type": "boolean"
          },
          "headers": {
            "items": {
              "$ref": "#/components/schemas/HeaderDto"
            },
            "type": "array"
          },
          "hstsEnabled": {
            "type": "boolean"
          },
          "loadBalancingMethod": {
            "type": "string"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationDto"
            },
            "type": "array"
          },
          "maxBodySizeBytes": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "nginxInstanceId": {
            "format": "uuid",
            "type": "string"
          },
          "readTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "sendTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "sslCertificateId": {
            "format": "uuid",
            "type": "string"
          },
          "sslEnabled": {
            "type": "boolean"
          },
          "upstreams": {
            "items": {
              "$ref": "#/components/schemas/UpstreamDto"
            },
            "minItems": 1,
            "type": "array"
          },
          "websocketEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "domain",
          "name",
          "nginxInstanceId",
          "upstreams"
        ],
        "type": "object"
      },
      "ProxySiteResponse": {
        "description": "One proxy site in full.",
        "properties": {
          "activeFrom": {
            "format": "date-time",
            "type": "string"
          },
          "connectTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "forceHttps": {
            "type": "boolean"
          },
          "headers": {
            "items": {
              "$ref": "#/components/schemas/HeaderDto"
            },
            "type": "array"
          },
          "hstsEnabled": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "loadBalancingMethod": {
            "type": "string"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationDto"
            },
            "type": "array"
          },
          "maxBodySizeBytes": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "nginxInstanceId": {
            "format": "uuid",
            "type": "string"
          },
          "readTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "secondsUntilExpiry": {
            "format": "int64",
            "type": "integer"
          },
          "sendTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "sslCertificateId": {
            "format": "uuid",
            "type": "string"
          },
          "sslEnabled": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "upstreams": {
            "items": {
              "$ref": "#/components/schemas/UpstreamDto"
            },
            "type": "array"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "websocketEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "connectTimeoutSeconds",
          "createdAt",
          "createdBy",
          "domain",
          "enabled",
          "forceHttps",
          "headers",
          "hstsEnabled",
          "id",
          "loadBalancingMethod",
          "locations",
          "maxBodySizeBytes",
          "name",
          "nginxInstanceId",
          "readTimeoutSeconds",
          "sendTimeoutSeconds",
          "sslEnabled",
          "status",
          "updatedAt",
          "updatedBy",
          "upstreams",
          "version",
          "websocketEnabled"
        ],
        "type": "object"
      },
      "ProxySiteSummaryResponse": {
        "description": "A proxy site, for listings.",
        "properties": {
          "activeFrom": {
            "format": "date-time",
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nginxInstanceId": {
            "format": "uuid",
            "type": "string"
          },
          "primaryUpstream": {
            "type": "string"
          },
          "secondsUntilExpiry": {
            "format": "int64",
            "type": "integer"
          },
          "sslEnabled": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "upstreamCount": {
            "format": "int32",
            "type": "integer"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "domain",
          "enabled",
          "id",
          "name",
          "nginxInstanceId",
          "sslEnabled",
          "status",
          "updatedAt",
          "upstreamCount",
          "version"
        ],
        "type": "object"
      },
...
      "RenewRequest": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "expiresAt"
        ],
        "type": "object"
      },
      "RequestAcmeRequest": {
        "properties": {
          "autoRenew": {
            "type": "boolean"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array"
          },
          "name": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "renewBeforeDays": {
            "format": "int32",
            "maximum": 89,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "domains",
          "name"
        ],
        "type": "object"
      },
...
      "UpstreamDto": {
        "description": "One upstream target.",
        "properties": {
          "backup": {
            "type": "boolean"
          },
          "failTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "host": {
            "minLength": 1,
            "type": "string"
          },
          "maxFails": {
            "format": "int32",
            "type": "integer"
          },
          "port": {
            "format": "int32",
            "maximum": 65535,
            "minimum": 1,
            "type": "integer"
          },
          "scheme": {
            "type": "string"
          },
          "weight": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "host",
          "port",
          "scheme"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer-jwt": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Management API for reverse-proxy sites across one or more NGINX hosts.\n\nAuthentication is OIDC: obtain an access token from Keycloak and send it as\na bearer token. Errors are RFC 9457 problem documents; branch on the stable\n`type` URI rather than on the human-readable `detail`.\n",
    "license": {
      "name": "Proprietary"
    },
    "title": "Easy NGINX Admin API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/certificates": {
      "get": {
        "description": "Metadata only. Private key material is never returned by any endpoint.",
        "operationId": "list_4",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CertificateResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List certificates",
        "tags": [
          "Certificates"
        ]
      },
      "post": {
        "description": "Validates with HTTP-01 against the managed NGINX hosts. Wildcards need DNS-01 and are refused. Requires MANAGE over the namespace of every domain requested, because a certificate is authority to terminate TLS for that name.",
        "operationId": "request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAcmeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Request a certificate over ACME",
        "tags": [
          "Certificates"
        ]
      }
    },
...
    "/api/v1/proxy-sites": {
      "get": {
        "description": "Filtering, sorting and pagination are applied in the query, never over an already-fetched page. Sort as sort=domain,asc; repeat the parameter for multiple sorts.",
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "nginxInstanceId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sslEnabled",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "expiringBefore",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 25,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseProxySiteSummaryResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List proxy sites",
        "tags": [
          "Proxy sites"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxySiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProxySiteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create a proxy site",
        "tags": [
          "Proxy sites"
        ]
      }
    },
    "/api/v1/proxy-sites/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Delete a proxy site",
        "tags": [
          "Proxy sites"
        ]
      },
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProxySiteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Fetch one proxy site",
        "tags": [
          "Proxy sites"
        ]
      },
      "put": {
        "description": "Send the version last read as If-Match to have concurrent edits rejected.",
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxySiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProxySiteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Replace a proxy site",
        "tags": [
          "Proxy sites"
        ]
      }
    },
    "/api/v1/proxy-sites/{id}/clone": {
      "post": {
        "description": "The copy is created disabled so a half-edited duplicate cannot take traffic.",
        "operationId": "clone",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProxySiteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Clone a proxy site",
        "tags": [
          "Proxy sites"
        ]
      }
    },
...
    "/api/v1/proxy-sites/{id}/disable": {
      "post": {
        "operationId": "disable",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProxySiteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable a proxy site",
        "tags": [
          "Proxy sites"
        ]
      }
    },
    "/api/v1/proxy-sites/{id}/enable": {
      "post": {
        "operationId": "enable",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProxySiteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Enable a proxy site",
        "tags": [
          "Proxy sites"
        ]
      }
    },
    "/api/v1/proxy-sites/{id}/expiration": {
      "delete": {
        "operationId": "removeExpiry",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProxySiteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Remove the expiry so the site never expires",
        "tags": [
          "Proxy sites"
        ]
      }
    },
...
    "/api/v1/proxy-sites/{id}/renew": {
      "post": {
        "operationId": "renew",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProxySiteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Extend or set the expiry",
        "tags": [
          "Proxy sites"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer-jwt": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
...
}